[HandshakeToAsync] Initial commit for HandshakeToAsync conversion
This commit implements a conversion between Handshake operations to Async operations. There is a rather natural mapping between the Async dialect and the Handshake dialect due to the asynchronous token-based flow that is a dataflow circuit.

Through this mapping, we have a way of converting a handshake circuit down to a multi-threaded (based on the implementation of the async runtime lib) compiled LLVMIR executable (performance not guaranteed ;).